+++ src/constants/index.ts
@@ -1,6 +1,7 @@
 export * from './actor-session';
 export * from './communication';
 export * from './cryptography';
+export * from './dataspace-discovery';
 export * from './dataspace-protocol';
 export * from './device';
 export * from './did-services';