+++ src/utils/index.ts
@@ -8,6 +8,7 @@ export * from './consent';
 export * from './did';
 export * from './did-resolution';
 export * from './dataspace-discovery';
+export * from './dataspace-discovery-defaults';
 export * from './dataspace-protocol';
 export * from './didcomm';
 export * from './didcomm-submit';